% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bistro.R
\name{bistro}
\alias{bistro}
\title{Identify human contributors to a bloodmeal using STR profiles}
\usage{
bistro(
  bloodmeal_profiles,
  human_profiles,
  kit,
  threshT,
  pop_allele_freqs = NULL,
  calc_allele_freqs = FALSE,
  bloodmeal_ids = NULL,
  human_ids = NULL,
  rm_twins = TRUE,
  modelDegrad = TRUE,
  modelBWstutt = FALSE,
  modelFWstutt = FALSE,
  difftol = 1,
  threads = 4,
  seed = 1,
  time_limit = 3
)
}
\arguments{
\item{bloodmeal_profiles}{Tibble with alleles for all bloodmeals in reference database including 4 columns: SampleName, Marker, Allele, Height. Height must be numeric or coercible to numeric.}

\item{human_profiles}{Tibble with alleles for all humans in reference database including three columns: SampleName, Marker, Allele.}

\item{kit}{STR kit name from euroformix. To see a list of all kits embedded in euroformix use \code{\link[euroformix:getKit]{euroformix::getKit()}}. If your kit is not included, see vignette("customized-kit") for details on how to include your own kit.}

\item{threshT}{Allele peak height threshold in RFUs. All peaks under this threshold will be filtered out. If prior filtering was performed, this number should be equal to or greater than that number.}

\item{pop_allele_freqs}{Data frame where the first column is the STR allele and the following columns are the frequency of that allele for different markers. Alleles that do not exist for a given marker are coded as NA. If NULL and \code{calc_allele_freqs = TRUE}, then population allele frequencies will be calculated from \code{human_profiles}.}

\item{calc_allele_freqs}{A boolean indicating whether or not to calculate allele frequencies from \code{human_profiles}. If FALSE, a \code{pop_allele_freqs} input is required. Default: FALSE}

\item{bloodmeal_ids}{Vector of bloodmeal ids from the SampleName column in \code{bloodmeal_profiles} for which to compute log10LRs. If NULL, all ids in the input dataframe will be used. Default: NULL}

\item{human_ids}{Vector of human ids from the SampleName column in \code{human_profiles} for which to compute log10LRs. If NULL, all ids in the input dataframe will be used. Default: NULL}

\item{rm_twins}{A boolean indicating whether or not to remove likely twins from the human database (identical STR profiles). Default: TRUE}

\item{modelDegrad}{A boolean indicating whether or not to model peak degradation. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: TRUE}

\item{modelBWstutt}{A boolean indicating whether or not to model peak backward stutter. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: FALSE}

\item{modelFWstutt}{A boolean indicating whether or not to model peak forward stutter. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: FALSE}

\item{difftol}{Tolerance for difference in log likelihoods across 2 iterations. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: 1}

\item{threads}{Number of threads to use when calculating log10LRs. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: 4}

\item{seed}{Seed when calculating log10LRs. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: 1}

\item{time_limit}{Time limit in minutes to run the \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} function on 1 bloodmeal-human pair. Default: 3}
}
\value{
Tibble with matches for bloodmeal-human pairs including the columns listed below. Note that if multiple matches are found for a bloodmeal, these are included as separate rows.
\itemize{
\item \code{bloodmeal_id}: bloodmeal id
\item \code{bloodmeal_locus_count}: number of loci successfully typed in the bloodmeal
\item \code{est_noc}: estimated number of contributors to the bloodmeal
\item \code{match}: whether or not a match was identified for a given bloodmeal (yes or no)
\item \code{human_id}: If match, human id (NA otherwise)
\item \code{log10LR}: If match, log10 likelihood ratio (NA otherwise)
\item \code{note}: Why the bloodmeal does or doesn't have a match
}
}
\description{
Identifies matches between bloodmeal STR profiles and a database of human STR profiles. The \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} function is used to calculate log10 likelihood ratios (log10LRs) that are then used to identify human contributors to each bloodmeal.
}
\examples{
bistro(bloodmeal_profiles, human_profiles,
pop_allele_freqs = pop_allele_freqs,
kit = "ESX17", threshT = 200)
}
