% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bloodmeal_profiles}
\alias{bloodmeal_profiles}
\title{Bloodmeal STR profiles}
\format{
\subsection{\code{bloodmeal_profiles}}{

A data frame with 69 rows and 4 columns:
\describe{
\item{SampleName}{Human sample identifier}
\item{Marker}{STR marker}
\item{Allele}{STR allele}
\item{Height}{STR peak height}
}
}
}
\source{
\url{http://euroformix.com/}
}
\usage{
bloodmeal_profiles
}
\description{
Example "bloodmeal" STR profiles from EuroForMix
}
\keyword{datasets}
