% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log10LRs.R
\name{calc_log10LRs}
\alias{calc_log10LRs}
\title{Calculate log10LRs for multiple bloodmeal-human pairs}
\usage{
calc_log10LRs(
  bloodmeal_profiles,
  human_profiles,
  pop_allele_freqs,
  kit,
  threshT,
  bloodmeal_ids = NULL,
  human_ids = NULL,
  modelDegrad = TRUE,
  modelBWstutt = FALSE,
  modelFWstutt = FALSE,
  difftol = 1,
  threads = 4,
  seed = 1,
  time_limit = 3
)
}
\arguments{
\item{bloodmeal_profiles}{Tibble with alleles for all bloodmeals in reference database including 4 columns: SampleName, Marker, Allele, Height. Height must be numeric or coercible to numeric.}

\item{human_profiles}{Tibble with alleles for all humans in reference database including three columns: SampleName, Marker, Allele.}

\item{pop_allele_freqs}{Data frame where the first column is the STR allele and the following columns are the frequency of that allele for different markers. Alleles that do not exist for a given marker are coded as NA. If NULL and \code{calc_allele_freqs = TRUE}, then population allele frequencies will be calculated from \code{human_profiles}.}

\item{kit}{STR kit name from euroformix. To see a list of all kits embedded in euroformix use \code{\link[euroformix:getKit]{euroformix::getKit()}}. If your kit is not included, see vignette("bistro") for details on how to include your own kit.}

\item{threshT}{Allele peak height threshold in RFUs. All peaks under this threshold will be filtered out. If prior filtering was performed, this number should be equal to or greater than that number.}

\item{bloodmeal_ids}{Vector of bloodmeal ids from the SampleName column in \code{bloodmeal_profiles} for which to compute log10LRs. If NULL, all ids in the input dataframe will be used. Default: NULL}

\item{human_ids}{Vector of human ids from the SampleName column in \code{human_profiles} for which to compute log10LRs. If NULL, all ids in the input dataframe will be used. Default: NULL}

\item{modelDegrad}{A boolean indicating whether or not to model peak degradation. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: TRUE}

\item{modelBWstutt}{A boolean indicating whether or not to model peak backward stutter. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: FALSE}

\item{modelFWstutt}{A boolean indicating whether or not to model peak forward stutter. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: FALSE}

\item{difftol}{Tolerance for difference in log likelihoods across 2 iterations. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: 1}

\item{threads}{Number of threads to use when calculating log10LRs. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: 4}

\item{seed}{Seed when calculating log10LRs. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: 1}

\item{time_limit}{Time limit in minutes to run the \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} function on 1 bloodmeal-human pair. Default: 3}
}
\value{
A tibble with the same output as for \code{\link[=bistro]{bistro()}}, except there is no match column and every bloodmeal-human pair with the calculated log10LR is included. The only additional column is \code{efm_noc} which is the number of contributors used for \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}}. This has a maximum value of 3.
}
\description{
Calculate log10LRs for multiple bloodmeal-human pairs
}
\keyword{internal}
