% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log10LRs.R
\name{calc_one_log10LR}
\alias{calc_one_log10LR}
\title{Calculate log10LR for one bloodmeal-human pair}
\usage{
calc_one_log10LR(
  bloodmeal_profiles,
  bloodmeal_id,
  human_profiles,
  human_id,
  pop_allele_freqs,
  kit,
  threshT,
  modelDegrad = TRUE,
  modelBWstutt = FALSE,
  modelFWstutt = FALSE,
  difftol = 1,
  threads = 4,
  seed = 1,
  time_limit = 3
)
}
\arguments{
\item{bloodmeal_profiles}{Tibble with alleles for all bloodmeals in reference database including 4 columns: SampleName, Marker, Allele, Height. Height must be numeric or coercible to numeric.}

\item{bloodmeal_id}{Bloodmeal id from the \code{SampleName} column in \code{bloodmeal_profiles} for which to compute the log10LR}

\item{human_profiles}{Tibble with alleles for all humans in reference database including three columns: SampleName, Marker, Allele.}

\item{human_id}{Human id from the \code{SampleName} column in \code{human_profiles} for which to compute the log10LR}

\item{pop_allele_freqs}{Data frame where the first column is the STR allele and the following columns are the frequency of that allele for different markers. Alleles that do not exist for a given marker are coded as NA. If NULL and \code{calc_allele_freqs = TRUE}, then population allele frequencies will be calculated from \code{human_profiles}.}

\item{kit}{STR kit name from euroformix. To see a list of all kits embedded in euroformix use \code{\link[euroformix:getKit]{euroformix::getKit()}}. If your kit is not included, see vignette("customized-kit") for details on how to include your own kit.}

\item{threshT}{Allele peak height threshold in RFUs. All peaks under this threshold will be filtered out. If prior filtering was performed, this number should be equal to or greater than that number.}

\item{modelDegrad}{A boolean indicating whether or not to model peak degradation. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: TRUE}

\item{modelBWstutt}{A boolean indicating whether or not to model peak backward stutter. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: FALSE}

\item{modelFWstutt}{A boolean indicating whether or not to model peak forward stutter. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: FALSE}

\item{difftol}{Tolerance for difference in log likelihoods across 2 iterations. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: 1}

\item{threads}{Number of threads to use when calculating log10LRs. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: 4}

\item{seed}{Seed when calculating log10LRs. \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} argument. Default: 1}

\item{time_limit}{Time limit in minutes to run the \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}} function on 1 bloodmeal-human pair. Default: 3}
}
\value{
tibble with log10LR for bloodmeal-human pair including bloodmeal_id, human_id, bloodmeal_locus_count (number of STR loci used for matching), est_noc (estimated number of contributors), efm_noc (number of contributors used in euroformix), log10LR (log10 likelihood ratio), note
}
\description{
Calculate log10LR for one bloodmeal-human pair
}
