% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_calc_allele_freqs}
\alias{check_calc_allele_freqs}
\title{Check calc_allele_freqs}
\usage{
check_calc_allele_freqs(calc_allele_freqs)
}
\arguments{
\item{calc_allele_freqs}{A boolean indicating whether or not to calculate allele frequencies from \code{human_profiles}. If FALSE, a \code{pop_allele_freqs} input is required. Default: FALSE}
}
\value{
Error or nothing
}
\description{
Check calc_allele_freqs
}
