% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_colnames}
\alias{check_colnames}
\title{Check column names}
\usage{
check_colnames(df, expected_colnames)
}
\arguments{
\item{df}{Dataframe to check}

\item{expected_colnames}{Column names required to be resent}
}
\value{
Nothing or error
}
\description{
Check column names
}
