% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_create_db_input}
\alias{check_create_db_input}
\title{Check input to `create_db_from_bloodmeals()``}
\usage{
check_create_db_input(bloodmeal_profiles, kit, peak_thresh, rm_markers)
}
\arguments{
\item{bloodmeal_profiles}{Tibble or data frame with alleles for all
bloodmeals in reference database including 4 columns: SampleName, Marker,
Allele, Height. Height must be numeric or coercible to numeric.}

\item{kit}{STR kit name from euroformix. To see a list of all kits embedded
in euroformix use \code{\link[euroformix:getKit]{euroformix::getKit()}}. If your kit is not included, see
vignette("bistro") for details on how to include your own kit.}

\item{peak_thresh}{Allele peak height threshold in RFUs. All peaks under this
threshold will be filtered out. If prior filtering was performed, this
number should be equal to or greater than that number. Also used for
\code{threshT} argument in \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}}.}

\item{rm_markers}{A vector indicating what markers should be removed prior to
calculating log10LRs. NULL to include all markers. By default, for the
bistro function AMEL is removed as it is not standard to include it in LR
calculations.}
}
\value{
number of alleles in a complete profile
}
\description{
Check input to `create_db_from_bloodmeals()``
}
