% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ids}
\alias{check_ids}
\title{Check input ids}
\usage{
check_ids(vec, vec_name)
}
\arguments{
\item{vec}{vector to check}

\item{vec_name}{vector name}
}
\value{
Error or nothing
}
\description{
Check input ids
}
\keyword{internal}
