% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_bool}
\alias{check_is_bool}
\title{Check is boolean}
\usage{
check_is_bool(vec, vec_name)
}
\arguments{
\item{vec}{vector to check}

\item{vec_name}{vector name}
}
\value{
Error or nothing
}
\description{
Check is boolean
}
\keyword{internal}
