% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_numeric}
\alias{check_is_numeric}
\title{Check is numeric}
\usage{
check_is_numeric(vec, pos = FALSE)
}
\arguments{
\item{vec}{vector to check}

\item{pos}{whether the number must be positive}

\item{vec_name}{vector name}
}
\value{
Error or nothing
}
\description{
Check is numeric
}
\keyword{internal}
