% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_kit}
\alias{check_kit}
\title{Check kit}
\usage{
check_kit(kit)
}
\arguments{
\item{kit}{STR kit name from euroformix. To see a list of all kits embedded
in euroformix use \code{\link[euroformix:getKit]{euroformix::getKit()}}. If your kit is not included, see
vignette("bistro") for details on how to include your own kit.}
}
\value{
Error if kit doesn't exist, dataframe of kit otherwise
}
\description{
Check kit
}
\keyword{internal}
