% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_peak_threshold}
\alias{check_peak_threshold}
\title{Check peak_threshold}
\usage{
check_peak_threshold(peak_threshold)
}
\arguments{
\item{peak_threshold}{Allele peak height threshold
in RFUs. All peaks under this
threshold will be filtered out. If prior filtering was performed, this
number should be equal to or greater than that number.
Also used for \code{threshT} argument in \code{\link[euroformix:contLikSearch]{euroformix::contLikSearch()}}.}
}
\value{
Error or nothing
}
\description{
Check peak_threshold
}
\keyword{internal}
