% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pkg_version}
\alias{check_pkg_version}
\title{Check package version}
\usage{
check_pkg_version(pkg, curr_version, version)
}
\arguments{
\item{pkg}{package to test}

\item{curr_version}{current package version}

\item{version}{required package version}
}
\value{
nothing or error if package version too old
}
\description{
Check package version
}
