% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_present}
\alias{check_present}
\title{Check if input markers to remove are present in the dataset}
\usage{
check_present(to_rm, df, col)
}
\arguments{
\item{df}{Data frame to check against}
}
\value{
Warning or nothing
}
\description{
Check if input markers to remove are present in the dataset
}
\keyword{internal}
