% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_setdiff_markers}
\alias{check_setdiff_markers}
\title{Check what markers are present}
\usage{
check_setdiff_markers(markers1, markers2, markers1_name, markers2_name)
}
\arguments{
\item{markers1}{first vector to check}

\item{markers2}{second vector to check}

\item{markers1_name}{first vector name}

\item{markers2_name}{second vector name}
}
\value{
Error or nothing
}
\description{
Check what markers are present
}
