% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_threshT}
\alias{check_threshT}
\title{Check threshT}
\usage{
check_threshT(threshT)
}
\arguments{
\item{threshT}{Allele peak height threshold in RFUs. All peaks under this threshold will be filtered out. If prior filtering was performed, this number should be equal to or greater than that number.}
}
\value{
Error or nothing
}
\description{
Check threshT
}
\keyword{internal}
