% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peaks.R
\name{filter_peaks}
\alias{filter_peaks}
\title{Filter peak heights that are under threshold}
\usage{
filter_peaks(bloodmeal_profiles, threshT)
}
\arguments{
\item{bloodmeal_profiles}{Tibble with alleles for all bloodmeals in reference database including 4 columns: SampleName, Marker, Allele, Height. Height must be numeric or coercible to numeric.}

\item{threshT}{Allele peak height threshold in RFUs. All peaks under this threshold will be filtered out. If prior filtering was performed, this number should be equal to or greater than that number.}
}
\value{
Filtered dataframe, error if all peak heights under threshold
}
\description{
Filter peak heights that are under threshold
}
