% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_bloodmeal_profiles}
\alias{format_bloodmeal_profiles}
\title{Format bloodmeal STR profiles}
\usage{
format_bloodmeal_profiles(bloodmeal_profiles)
}
\arguments{
\item{bloodmeal_profiles}{Tibble or data frame with alleles for all
bloodmeals in reference database including 4 columns: SampleName, Marker,
Allele, Height. Height must be numeric or coercible to numeric.}
}
\value{
A list of lists including the alleles for each sample in the format
required for input into the euroformix \code{contLikSearch()} function.
}
\description{
Format bloodmeal STR profiles for input into the euroformix
\code{contLikSearch()} function.
}
\keyword{internal}
