% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_similarity.R
\name{get_bloodmeal_human_similarity}
\alias{get_bloodmeal_human_similarity}
\title{Identify matches between bloodmeal and human STR profiles
based on similarity values}
\usage{
get_bloodmeal_human_similarity(bloodmeal_profiles, human_profiles)
}
\arguments{
\item{bloodmeal_profiles}{Tibble with alleles for all bloodmeals in reference
database including 4 columns: SampleName, Marker, Allele, Height. Height
must be numeric or coercible to numeric.}

\item{human_profiles}{Tibble with alleles for all humans in reference
database including three columns: SampleName, Marker, Allele.}
}
\value{
dataframe with two columns:
\itemize{
\item bloodmeal_id: bloodmeal ID
\item human_id: human ID of match, or "No match" if no match
}
}
\description{
Identify matches between bloodmeal and human STR profiles
based on similarity values
}
\keyword{internal}
