% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_similarity.R
\name{get_human_similarity}
\alias{get_human_similarity}
\title{Get pairwise human STR profile similarities}
\usage{
get_human_similarity(human_profiles)
}
\arguments{
\item{human_profiles}{Tibble with alleles for all humans in reference
database including three columns: SampleName, Marker, Allele.}
}
\value{
Dataframe with three columns:
\itemize{
\item hu1: human ID 1
\item hu2: human ID 2
\item similarity: similarity value (# exact locus matches / # loci)
}
}
\description{
Get pairwise human STR profile similarities
}
