% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_profiles}
\alias{human_profiles}
\title{Human reference database}
\format{
\subsection{\code{human_profiles}}{

A data frame with 96 rows and 3 columns:
\describe{
\item{SampleName}{Human sample identifier}
\item{Marker}{STR marker}
\item{Allele}{STR allele}
}
}
}
\source{
\url{http://euroformix.com/}
}
\usage{
human_profiles
}
\description{
Example human STR profiles from EuroForMix
}
\keyword{datasets}
