% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_matches.R
\name{identify_matches}
\alias{identify_matches}
\title{Identify matches for multiple bloodmeal-human pairs}
\usage{
identify_matches(log10LRs, bloodmeal_ids = NULL)
}
\arguments{
\item{log10LRs}{Output from \code{\link[=calc_log10LRs]{calc_log10LRs()}}}

\item{bloodmeal_ids}{Vector of bloodmeal ids from the SampleName column in \code{bloodmeal_profiles} for which to compute log10LRs. If NULL, all ids in the input dataframe will be used. Default: NULL}
}
\value{
A tibble with the same output as for \code{\link[=bistro]{bistro()}}.
}
\description{
Identify matches for multiple bloodmeal-human pairs
}
\keyword{internal}
