% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_matches.R
\name{identify_one_match_set}
\alias{identify_one_match_set}
\title{Identify matches between a bloodmeal and human references}
\usage{
identify_one_match_set(log10_lrs, bloodmeal_id)
}
\arguments{
\item{log10_lrs}{output from \code{\link[=calc_log10_lrs]{calc_log10_lrs()}}}

\item{bloodmeal_id}{Bloodmeal id from the \code{SampleName} column in
\code{bloodmeal_profiles} for which to compute the log10_lr}
}
\value{
tibble with matches for bloodmeal-human pairs including bloodmeal_id,
locus_count (number of STR loci used for matching), est_noc
(estimated number of contributors), match, human_id (if match), log10_lr
(log10 likelihood ratio), notes
}
\description{
Identify matches between a bloodmeal and human references
}
\keyword{internal}
