% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_static_thresh.R
\name{match_static_thresh}
\alias{match_static_thresh}
\title{Match bloodmeals and humans based on log10LR static threshold}
\usage{
match_static_thresh(log10_lrs, thresh)
}
\arguments{
\item{log10_lrs}{Output from \code{\link[=calc_log10_lrs]{calc_log10_lrs()}} or from \code{bistro} with
\code{return_lrs = TRUE} (\code{lrs}: the second element in the list)}

\item{thresh}{log10LR threshold for matching.
All bloodmeal-human pairs with
a log10LR ≥ thresh will be considered a match.}
}
\value{
Dataframe including columns similar to the \code{bistro()} output.
}
\description{
Match bloodmeals and humans based on log10LR static threshold
}
\examples{
bistro_output <- bistro(bloodmeal_profiles, human_profiles,
  pop_allele_freqs = pop_allele_freqs,
  kit = "ESX17", peak_thresh = 200, return_lrs = TRUE
)
match_static_thresh(bistro_output$lrs, 10)
}
