% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_allele_freqs}
\alias{pop_allele_freqs}
\title{Human population allele frequencies}
\format{
\subsection{\code{pop_allele_freqs} A data frame with 76 rows and 17 columns. The}{

first column is the STR allele and the following columns are the frequency
of that allele for different markers.
}
}
\source{
\url{http://euroformix.com/}
}
\usage{
pop_allele_freqs
}
\description{
Example population allele frequencies from EuroForMix (ESX17_Norway)
}
\keyword{datasets}
