% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_dups}
\alias{rm_dups}
\title{Remove duplicate rows with warning}
\usage{
rm_dups(df)
}
\arguments{
\item{df}{Dataframe from which to remove duplicate rows}
}
\value{
Un-duplicated dataframe
}
\description{
Remove duplicate rows with warning
}
