% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{rm_markers}
\alias{rm_markers}
\title{Remove markers}
\usage{
rm_markers(profiles, markers)
}
\arguments{
\item{df}{Dataframe from which to remove markers}
}
\value{
Dataframe without markers
}
\description{
Remove markers
}
\keyword{internal}
