% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{rm_twins}
\alias{rm_twins}
\title{Remove identical STR profiles}
\usage{
rm_twins(human_profiles)
}
\arguments{
\item{human_profiles}{Tibble with alleles for all humans in reference
database including three columns: SampleName, Marker, Allele.}
}
\value{
Data frame with twins removed
}
\description{
Remove identical STR profiles
}
\keyword{internal}
