% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_ids}
\alias{subset_ids}
\title{Subset to ids present in the dataset}
\usage{
subset_ids(ids, vec, vec_name)
}
\arguments{
\item{ids}{ids to check if in vec}

\item{vec}{vector of ids}

\item{vec_name}{name of vector}
}
\value{
list of IDs that are present
}
\description{
Subset to ids present in the dataset
}
\keyword{internal}
