% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_allele_freqs.R
\name{calc_allele_freqs}
\alias{calc_allele_freqs}
\title{Calculate allele frequencies}
\usage{
calc_allele_freqs(human_profiles, rm_twins = TRUE)
}
\arguments{
\item{human_profiles}{Tibble with alleles for all humans in reference
database including three columns: SampleName, Marker, Allele.}

\item{rm_twins}{A boolean indicating whether or not to remove likely twins
from the human database (identical STR profiles). Default: TRUE}
}
\value{
A tibble where the first column is the STR allele and the following
columns are the frequency of that allele for different markers. Alleles
that do not exist for a given marker are coded as NA.
}
\description{
Calculate allele frequencies for a (generally human) population.
}
\keyword{internal}
