% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_matches.R
\name{identify_one_match_set}
\alias{identify_one_match_set}
\title{Identify matches between a bloodmeal and human references}
\usage{
identify_one_match_set(log10LRs, bloodmeal_id)
}
\arguments{
\item{log10LRs}{output from \code{\link[=calc_log10LRs]{calc_log10LRs()}}}

\item{bloodmeal_id}{Bloodmeal id from the \code{SampleName} column in \code{bloodmeal_profiles} for which to compute the log10LR}
}
\value{
tibble with matches for bloodmeal-human pairs including bloodmeal_id, bloodmeal_locus_count (number of STR loci used for matching), est_noc (estimated number of contributors), match, human_id (if match), log10LR (log10 likelihood ratio), note
}
\description{
Identify matches between a bloodmeal and human references
}
