% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{prep_human_profiles}
\alias{prep_human_profiles}
\title{Preprocess human profiles}
\usage{
prep_human_profiles(human_profiles, human_ids = NULL, rm_twins = TRUE)
}
\arguments{
\item{human_profiles}{Tibble with alleles for all humans in reference
database including three columns: SampleName, Marker, Allele.}

\item{human_ids}{Vector of human ids from the SampleName column in
\code{human_profiles} for which to compute log10_lrs. If NULL, all ids in the
input dataframe will be used. Default: NULL}

\item{rm_twins}{A boolean indicating whether or not to remove likely twins
(identical STR profiles) from the human database prior to identifying
matches. Default: TRUE}
}
\value{
Dataframe with preprocessed human profiles
}
\description{
Removes duplicates and optionally twins, subsets ids
}
\keyword{internal}
